+++ cvxpy/transforms/partial_optimize.py
@@ -83,9 +83,19 @@ class PartialProblem(Expression):
     def __init__(self, prob, opt_vars, dont_opt_vars):
         self.opt_vars = opt_vars
         self.dont_opt_vars = dont_opt_vars
-        self.args = [prob]
+        # Replace the opt_vars in prob with new variables.
+        id_to_new_var = {var.id:var.copy() for var in self.opt_vars}
+        new_obj = self._replace_new_vars(prob.objective, id_to_new_var)
+        new_constrs = [self._replace_new_vars(con, id_to_new_var)
+                       for con in prob.constraints]
+        self.args = [Problem(new_obj, new_constrs)]
         self.init_dcp_attr()
         super(PartialProblem, self).__init__()
+
+    def get_data(self):
+        """Returns info needed to reconstruct the expression besides the args.
+        """
+        return [self.opt_vars, self.dont_opt_vars]
 
     def init_dcp_attr(self):
         """Determines the curvature, sign, and shape from the arguments.
@@ -134,7 +144,7 @@ class PartialProblem(Expression):
                 return None
             else:
                 fix_vars += [var == var.value]
-        prob = Problem(type(self.args[0].objective)(self), fix_vars)
+        prob = Problem(self.args[0].objective.copy(self), fix_vars)
         result = prob.solve()
         # Restore the original values to the variables.
         for var in self.variables():
@@ -142,30 +152,34 @@ class PartialProblem(Expression):
         return result
 
     @staticmethod
-    def _replace_new_vars(expr, id_to_new_var):
-        """Replaces the given variables in the expression.
+    def _replace_new_vars(obj, id_to_new_var):
+        """Replaces the given variables in the object.
 
         Parameters
         ----------
-        expr : LinOp
-            The expression to replace variables in.
+        obj : Object
+            The object to replace variables in.
         id_to_new_var : dict
             A map of id to new variable.
 
         Returns
         -------
-        LinOp
-            An LinOp identical to expr, but with the given variables replaced.
+        Object
+            An object identical to obj, but with the given variables replaced.
         """
-        if expr.type == lo.VARIABLE and expr.data in id_to_new_var:
-            return id_to_new_var[expr.data]
+        if isinstance(obj, Variable) and obj.id in id_to_new_var:
+            return id_to_new_var[obj.id]
+        # Leaves outside of optimized variables are preserved.
+        elif len(obj.args) == 0:
+            return obj
+        # Parent nodes are copied.
         else:
             new_args = []
-            for arg in expr.args:
+            for arg in obj.args:
                 new_args.append(
                     PartialProblem._replace_new_vars(arg, id_to_new_var)
                 )
-            return lo.LinOp(expr.type, expr.size, new_args, expr.data)
+            return obj.copy(new_args)
 
     def canonicalize(self):
         """Returns the graph implementation of the object.
@@ -176,11 +190,4 @@ class PartialProblem(Expression):
         -------
             A tuple of (affine expression, [constraints]).
         """
-        id_to_new_var = {v.id:lu.create_var(v.size) for v in self.opt_vars}
-        obj, constr = self.args[0].canonical_form
-        obj = self._replace_new_vars(obj, id_to_new_var)
-        new_constr = []
-        for con in constr:
-            expr = self._replace_new_vars(con.expr, id_to_new_var)
-            new_constr += [type(con)(expr, con.constr_id, con.size)]
-        return obj, new_constr
+        return self.args[0].canonical_form